backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
openpyxl==3.1.2
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
django-extensions==3.2.3

